6.3.1

+++ src/main/java/InheritanceAndPolymorphism/figures/Square.java
@@ -1,22 +1,22 @@
 package InheritanceAndPolymorphism.figures;
 
-public class Square extends Rectangle{
-    public Square(int width, int height) {
+public class Square extends Rectangle {
+    public Square(double width, double height) {
         super(width, height);
     }
 
-    public Square(int width) {
+    public Square(double width) {
         super(width , width);
     }
 
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         setWidth(width);
         setHeight(width);
     }
 
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         setHeight(height);
         setWidth(height);
     }
